requirements.txt
peewee==3.2.2
pytest==3.5.0
python-decouple==3.1
